Move txn to own file
