+++ host-go/repository/registry.go
@@ -19,37 +19,6 @@ import (
 // the users if broken (deadlocks, large performance degradation).  It should have proper unit tests.
 // https://github.com/sourcenetwork/defradb/issues/1596
 
-// TxnSource represents an object capable of constructing the transactions that
-// implicit-transaction registries need internally.
-type TxnSource interface {
-	NewTxn(context.Context, bool) (Txn, error)
-}
-
-// Txn is a common interface to the BasicTxn struct.
-type Txn interface {
-	// ID returns the unique immutable identifier for this transaction.
-	ID() uint64
-
-	// Commit finalizes a transaction, attempting to commit it to the Datastore.
-	// May return an error if the transaction has gone stale. The presence of an
-	// error is an indication that the data was not committed to the Datastore.
-	Commit(ctx context.Context) error
-	// Discard throws away changes recorded in a transaction without committing
-	// them to the underlying Datastore. Any calls made to Discard after Commit
-	// has been successfully called will have no effect on the transaction and
-	// state of the Datastore, making it safe to defer.
-	Discard(ctx context.Context)
-
-	// OnSuccess registers a function to be called when the transaction is committed.
-	OnSuccess(fn func())
-
-	// OnError registers a function to be called when the transaction is rolled back.
-	OnError(fn func())
-
-	// OnDiscard registers a function to be called when the transaction is discarded.
-	OnDiscard(fn func())
-}
-
 // LensRegistry exposes several useful thread-safe migration related functions which may
 // be used to manage migrations.
 type LensRegistry interface {
